backend/requirements/base.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
djangorestframework>=3.15,<4.0
django-environ>=0.11,<1.0
drf-spectacular>=0.27,<1.0
django-cors-headers>=4.4,<5.0
Pillow>=10.0,<13.0
gunicorn>=22.0,<24.0
djangorestframework-simplejwt[crypto]>=5.3,<6.0
django-jazzmin>=3.0,<4.0
psycopg[binary]
